Change ip neigh command to use replace instead of add
Pinging between h1a and h1b before the permanent neighbor entries are added results in a FAILED neighbor entry. This entry is not overwritten by "ip neigh add", resulting in the tutorial failing. Use "ip neigh replace" instead, which is safe for both no entries or a failed entry in the neighbor table.