Workaround for segmentrouting netcfg bug
Before we where pushing empty "interfaces" block as part of the
pre-solution netcfg-sr.json, but when applying the solution (with the
actual "interfaces" block for h3 and h4), segmentrouting was not
updating the ingress_inport_vlan table properly. For example, even if
interface for h3 was pushed as a tagged one (VLAN 300) the corresponding
rule was missing from the table.

In this workaorund, we avoid pushing the entire port block, which is
instead provided in the solution. segmnetrouting seems to be happy.