Add EXERCISE 8 on GTP termination with fabric.p4 (#47)
* Use custom stratum_bmv2 image for both mininet and ptf

This is required to use scapy in Mininet. Before the ptf image was
essentially mn-stratum plus scapy, ptf, etc. I renamed that to a generic
stratum_bmv2 and use it for both mininet  (in docker-compose) and to run
ptf tests, instead of using two different images.

* Progress on exercise 8

* Bump stratum_bmv2

* Fix loop by disabling ip4 forwarding

Perhaps it is enabled by default in Docker for Mac?

* Fix flowrule-gtp.json solution

* Add make targets to push flow rules

* Check gtp solution in Travis

* Add instructions for Exercise 8

* Not exercise 4

* Update topo img

* Edits

* Add link to main readme